requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.6.1
openpyxl==3.1.5
pandas==2.2.3
PyYAML==6.0.2
python-multipart==0.0.12
aiofiles==24.1.0
anthropic==0.40.0
pywin32==308; sys_platform == "win32"
extract-msg==0.52.0
python-dotenv==1.0.1
openai==1.58.1
